fix the review comments: moving stats from connection_base to connection & connection gauage at peer
