+++ node/connection_base.js
@@ -31,7 +31,6 @@ var Operations = require('./operations');
 
 var DEFAULT_OUTGOING_REQ_TIMEOUT = 2000;
 var CONNECTION_BASE_IDENTIFIER = 0;
-var DEFAULT_REPORT_INTERVAL = 1000;
 
 function TChannelConnectionBase(channel, direction, remoteAddr) {
     assert(!channel.destroyed, 'refuse to create connection for destroyed channel');
@@ -54,7 +53,6 @@ function TChannelConnectionBase(channel, direction, remoteAddr) {
     self.timers = channel.timers;
     self.direction = direction;
     self.remoteAddr = remoteAddr;
-    self.reportInterval = self.options.reportInterval || DEFAULT_REPORT_INTERVAL;
     self.timer = null;
     self.remoteName = null; // filled in by identify message
 
@@ -73,151 +71,11 @@ function TChannelConnectionBase(channel, direction, remoteAddr) {
 
     self.tracer = self.channel.tracer;
     self.ops.startTimeoutTimer();
-
-    if (direction === 'out') {
-        self.channel.connectionsInitiatedStat.increment(1, {
-            'host-port': self.channel.hostPort || '0.0.0.0:0',
-            'peer-host-port': self.remoteAddr
-        });
-    } else {
-        self.channel.connectionsAcceptedStat.increment(1, {
-            'host-port': self.channel.hostPort,
-            'peer-host-port': self.remoteAddr
-        });
-    }
-
-    if (self.reportInterval > 0) {
-        self.reportTimer = self.timers.setTimeout(
-            onReport, self.reportInterval
-        );
-    }
-
-    function onReport() {
-        self.channel.connectionsActiveStat.update(1, {
-            'host-port': self.channel.hostPort,
-            'peer-host-port': self.remoteAddr
-        });
-    }
 }
 inherits(TChannelConnectionBase, EventEmitter);
 
 TChannelConnectionBase.prototype.close = function close(callback) {
-    var self = this;
-
-    self.resetAll(errors.LocalSocketCloseError());
     callback();
-};
-
-// this connection is completely broken, and is going away
-// In addition to erroring out all of the pending work, we reset the state in case anybody
-// stumbles across this object in a core dump.
-TChannelConnectionBase.prototype.resetAll = function resetAll(err) {
-    var self = this;
-
-    self.ops.destroy();
-
-    if (self.closing) return;
-    self.closing = true;
-    self.closeError = err;
-    self.closeEvent.emit(self, err);
-
-    var requests = self.ops.getRequests();
-    var pending = self.ops.getPending();
-
-    var inOpKeys = Object.keys(requests.in);
-    var outOpKeys = Object.keys(requests.out);
-
-    if (!err) {
-        err = new Error('unknown connection reset'); // TODO typed error
-    }
-
-    if (!self.remoteName) {
-        if (self.direction === 'out') {
-            self.channel.connectionsConnectErrorsStat.increment(1, {
-                'host-port': self.channel.hostPort || '0.0.0.0:0',
-                'peer-host-port': self.remoteAddr
-            });
-        } else {
-            self.channel.connectionsAcceptedErrorsStat.increment(1, {
-                'host-port': self.channel.hostPort
-            });
-        }
-    } else {
-        if (err.type !== 'tchannel.socket-local-closed' &&
-            err.type !== 'tchannel.socket-closed') {
-            self.channel.connectionsErrorsStat.increment(1, {
-                'host-port': self.channel.hostPort || '0.0.0.0:0',
-                'peer-host-port': self.remoteAddr,
-                'type': err.type // TODO unified error type
-            });
-        }
-
-        self.channel.connectionsClosedStat.increment(1, {
-            'host-port': self.channel.hostPort || '0.0.0.0:0',
-            'peer-host-port': self.remoteAddr,
-            'reason': err.type // TODO unified reason type
-        });
-    }
-
-    if (self.reportTimer) {
-        self.timers.clearTimeout(self.reportTimer);
-        self.reportTimer = null;
-    }
-
-    var logInfo = {
-        error: err,
-        remoteName: self.remoteName,
-        localName: self.channel.hostPort,
-        numInOps: inOpKeys.length,
-        numOutOps: outOpKeys.length,
-        inPending: pending.in,
-        outPending: pending.out
-    };
-
-    if (err.type && err.type.lastIndexOf('tchannel.socket', 0) < 0) {
-        self.logger.warn('resetting connection', logInfo);
-        self.errorEvent.emit(self, err);
-    } else if (
-        err.type !== 'tchannel.socket-closed' &&
-        err.type !== 'tchannel.socket-local-closed'
-    ) {
-        logInfo.error = extend(err);
-        logInfo.error.message = err.message;
-        self.logger.info('resetting connection', logInfo);
-    }
-
-    // requests that we've received we can delete, but these reqs may have started their
-    //   own outgoing work, which is hard to cancel. By setting this.closing, we make sure
-    //   that once they do finish that their callback will swallow the response.
-    inOpKeys.forEach(function eachInOp(id) {
-        // TODO: support canceling pending handlers
-        self.ops.removeReq(id);
-        // TODO report or handle or log errors or something
-    });
-
-    // for all outgoing requests, forward the triggering error to the user callback
-    outOpKeys.forEach(function eachOutOp(id) {
-        var req = requests.out[id];
-        self.ops.removeReq(id);
-
-        var info = {
-            remoteAddr: self.remoteAddr,
-            direction: self.direction,
-            remoteName: self.remoteName,
-            reqRemoteAddr: req.remoteAddr,
-            serviceName: req.serviceName,
-            outArg1: String(req.arg1)
-        };
-        if (err.type === 'tchannel.socket-local-closed') {
-            err = errors.TChannelLocalResetError(err, info);
-        } else {
-            err = errors.TChannelConnectionResetError(err, info);
-        }
-
-        req.errorEvent.emit(req, err);
-    });
-
-    self.ops.clear();
 };
 
 // create a request
